CSE499B/Final project/cse499b_inpainting_deploy/requirements.txt
gradio>=4.44.0
pillow>=10.0.0
numpy>=1.26.0
torch>=2.2.0
torchvision>=0.17.0
diffusers>=0.30.0
transformers>=4.44.0
accelerate>=0.33.0
safetensors>=0.4.0
